+++ test-parts/part-e-rio_settings/rio-settings-main.js
@@ -1,19 +1,16 @@
-const cRemoteIoSettingsFile = require("./settings/set-main");
-const cRemoteIoPropertiesFile = require("./settings/set-props");
-const cRemoteIoFunctionsFile = require("./settings/set-functions");
+const mainTests = require("./settings/set-main");
+const propertyTests = require("./settings/set-props");
+const functionTests = require("./settings/set-functions");
 
 
 function coordinateSettings()
 {
 	describe("E - Remote IO Settings", function()
 	{
-		cRemoteIoSettingsFile.callTestRemoteIoSettings();
-		cRemoteIoPropertiesFile.callTestRemoteIoProperties();
-		cRemoteIoFunctionsFile.callTestRemoteIoFunctions();
+		mainTests();
+		propertyTests();
+		functionTests();
 	});
 }
 
-module.exports =
-{
-	callSettings: coordinateSettings
-}
+module.exports = coordinateSettings;