+++ test-parts/part-e-rio_settings/settings/set-main.js
@@ -68,7 +68,4 @@ function checkExports()
 	});
 }
 
-module.exports =
-{
-	callTestRemoteIoSettings: testRemoteIoSettings
-};
+module.exports = testRemoteIoSettings;